prioritized_metered_channel::oneshot

Trait Measurable

pub trait Measurable {
    // Required method
    fn measurements(&self) -> Measurements;
}

A dummy trait that allows implementing measurements for Result<_,_>.

Required Methods

fn measurements(&self) -> Measurements

Obtain a set of measurements represented by the Measurements type.