Struct prioritized_metered_channel::CoarseDuration

pub struct CoarseDuration(_);

A duration type to represent an approximate span of time

Implementations

impl Duration

pub fn new(sec: u64, nanos: u32) -> Duration

Creates a new Duration from the specified number of seconds and additional nanosecond precision

pub fn from_days(days: u64) -> Duration

Creates a new Duration from the specified number of days

pub fn from_hours(hours: u64) -> Duration

Creates a new Duration from the specified number of hours

pub fn from_mins(mins: u64) -> Duration

Creates a new Duration from the specified number of minutes

pub fn from_secs(secs: u64) -> Duration

Creates a new Duration from the specified number of seconds

pub fn from_millis(millis: u64) -> Duration

Creates a new Duration from the specified number of milliseconds

pub fn as_days(&self) -> u64

Returns the number of days represented by this duration

pub fn as_hours(&self) -> u64

Returns the number of minutes represented by this duration

pub fn as_mins(&self) -> u64

Returns the number of minutes represented by this duration

pub fn as_secs(&self) -> u64

Returns the number of whole seconds represented by this duration

pub fn as_millis(&self) -> u64

Returns the number of whole milliseconds represented by this duration

pub fn as_micros(&self) -> u64

Returns the number of whole microseconds represented by this duration

pub fn as_nanos(&self) -> u64

Returns the number of whole nanoseconds represented by this duration

pub fn subsec_nanos(&self) -> u32

Returns the nanosecond precision represented by this duration

pub fn as_ticks(&self) -> u64

Return this duration as a number of “ticks”.

Note that length of a ‘tick’ is not guaranteed to represent the same amount of time across different platforms, or from one version of coarsetime to another.

pub fn from_ticks(ticks: u64) -> Duration

Creates a new Duration from the specified number of “ticks”.

Note that length of a ‘tick’ is not guaranteed to represent the same amount of time across different platforms, or from one version of coarsetime to another.

pub fn as_f64(&self) -> f64

Returns the duration as a floating point number, representing the number of seconds