Struct prioritized_metered_channel::MeteredSender

pub struct MeteredSender<T> { /* private fields */ }

The sender component, tracking the number of items sent across it.

Implementations

impl<T> MeteredSender<T>

pub fn meter(&self) -> &Meter

Get an updated accessor object for all metrics collected.

pub async fn send(&mut self, msg: T) -> Result<(), SendError>where
Self: Unpin,

Send message, wait until capacity is available.

pub fn try_send(
&mut self,
msg: T
) -> Result<(), TrySendError<MaybeTimeOfFlight<T>>>

Attempt to send message or fail immediately.