Struct prioritized_metered_channel::UnboundedMeteredSender

pub struct UnboundedMeteredSender<T> { /* private fields */ }

The sender component, tracking the number of items sent across it.

Implementations

impl<T> UnboundedMeteredSender<T>

pub fn meter(&self) -> &Meter

Get an updated accessor object for all metrics collected.

pub fn unbounded_send( &self, msg: T, ) -> Result<(), TrySendError<MaybeTimeOfFlight<T>>>

Attempt to send message or fail immediately.