prioritized_metered_channel

Module oneshot

Metered variant of oneshot channels to be able to extract delays caused by delayed responses.

Structs

Measurements

Obtained measurements by the Receiver side of the MeteredOneshot.

MeteredReceiver

Oneshot receiver, created by channel.

MeteredSender

Oneshot sender, created by channel.

OutputWithMeasurements

A wrapping type for the actual type T that is sent with the oneshot yet allow to attach Measurements to it.

Enums

Error

Reason

Provides the reason for termination.

Traits

Measurable

A dummy trait that allows implementing measurements for Result<_,_>.

Functions

channel

Create a new pair of OneshotMetered{Sender,Receiver}.